pyproject.toml
[tool.poetry]
name = "bobolink"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.6"
django-extensions = "^2.2.9"
django-debug-toolbar = "^2.2"
werkzeug = "^1.0.1"
django-environ = "^0.4.5"
djangorestframework = "^3.11.0"
django-filter = "^2.3.0"
djoser = "^2.0.3"
django-taggit = "^1.3.0"
django-ordered-model = "^3.4.1"

[tool.poetry.dev-dependencies]
prospector = "^1.3.0"
yapf = "^0.30.0"
mypy = "^0.782"
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
